Rename display-grid.tsx to stencil-display-grid.tsx
https://coveord.atlassian.net/browse/KIT-4200
