+++ packages/atomic/src/components/commerce/atomic-commerce-product-list/atomic-commerce-product-list.tsx
@@ -28,7 +28,6 @@ import {FocusTargetController} from '../../../utils/stencil-accessibility-utils'
 import {randomID} from '../../../utils/stencil-utils';
 import {ResultsPlaceholdersGuard} from '../../common/atomic-result-placeholder/placeholders';
 import {createAppLoadedListener} from '../../common/interface/store';
-import {DisplayGrid} from '../../common/item-list/display-grid';
 import {
   DisplayTable,
   DisplayTableData,
@@ -36,6 +35,7 @@ import {
 } from '../../common/item-list/display-table';
 import {DisplayWrapper} from '../../common/item-list/display-wrapper';
 import {ItemListGuard} from '../../common/item-list/item-list-guard';
+import {DisplayGrid} from '../../common/item-list/stencil-display-grid';
 import {
   ItemListCommon,
   ItemRenderingFunction,
